refactor: rename pages components
